+++ CHANGELOG.rst
@@ -15,7 +15,7 @@ Unreleased
 ~~~~~~~~~~
 Added
 -----
-* Adds search script datadoc_search.py, for searching Datadog monitors and dashboards.
+* Adds search script datadog_search.py, for searching Datadog monitors and dashboards.
 
 [5.1.0] - 2024-11-21
 ~~~~~~~~~~~~~~~~~~~~